requirements.txt
openai
requests
torch
numpy
Pillow